fix: mapping url 수정
