+++ auth-service/src/main/java/com/sobok/authservice/auth/controller/AuthController.java
@@ -122,15 +122,4 @@ public class AuthController {
 
     }
 
-    /**
-     * 라이더 활성화
-     */
-    @PutMapping("/active-rider")
-    public ResponseEntity<Void> activeRider(@RequestParam Long authId) {
-        authService.activeRider(authId);
-        return ResponseEntity.ok().build();
-    }
-
-
-
 }
